fix: fixing bugs

+++ Gas.Application/Features/AccessorySaleFeatures/Validator/AddAccessorySaleValidator.cs
@@ -21,7 +21,7 @@ namespace Gas.Application.Features.AccessorySaleFeatures.Validator
         {
             return number > 0;
         }
-        private bool BeValidFloat(float? number)
+        private bool BeValidFloat(double? number)
         {
             return number > 0;
         }
